+++ src/Resources/contao/languages/de/tl_news.php
@@ -1,4 +1,4 @@
 <?php
 
 $GLOBALS['TL_LANG']['tl_news']['dsolnews_legend'] = 'Zuordnung DSOL-Website';
-$GLOBALS['TL_LANG']['tl_news']['dsol_id'] = array('DSOL-ID', 'ID der Nachricht auf der DSOL-Website');
+$GLOBALS['TL_LANG']['tl_news']['dsol_id'] = array('DSOL-ID', 'Bitte nicht ändern! Die ID verbindet diese Nachricht mit der entsprechenden Nachricht auf der DSOL-Website. Wenn Sie hier 0 eintragen, wird auf der DSOL-Website eine neue Nachricht erzeugt. Bei anderen Werten, die nicht vom System eingetragen wurden, kann die Synchronisierung durcheinanderkommen.');