+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # DSOL-Nachrichten Changelog
 
+## Version 1.0.1 (2022-01-11)
+
+* Add: Hinweis für Änderungen der Spalte tl_news.dsol_id
+* Fix: Ersetzung der Inserttags im Teaser vor der Synchronisation
+* Add: Hinzufügen der Domain schachbund.de in Links des Teasers
+
 ## Version 1.0.0 (2022-01-11)
 
 * Aus- und Umbau des Bundles: Nicht in externe Tabelle content schreiben, sondern in eigene Tabelle dsb_content